move function from cairo file
