+++ stwo_cairo_prover/crates/prover/src/witness/public_data.rs
@@ -1,0 +1,121 @@
+use std::array;
+
+use cairo_air::public_data::{
+    MemorySmallValue, PublicData, PublicMemory, PublicSegmentRanges, SegmentRange,
+};
+use itertools::Itertools;
+use stwo_cairo_adapter::memory::Memory;
+use stwo_cairo_adapter::PublicSegmentContext;
+use stwo_cairo_common::prover_types::cpu::CasmState;
+
+pub fn extract_public_segments(
+    memory: &Memory,
+    initial_ap: u32,
+    final_ap: u32,
+    public_segment_context: PublicSegmentContext,
+) -> PublicSegmentRanges {
+    let n_public_segments = public_segment_context.iter().filter(|&b| *b).count() as u32;
+
+    let to_memory_value = |addr: u32| {
+        let id = memory.get_raw_id(addr);
+        let value = memory.get(addr).as_small() as u32;
+        MemorySmallValue { id, value }
+    };
+
+    let start_ptrs = (initial_ap..initial_ap + n_public_segments).map(to_memory_value);
+    let end_ptrs = (final_ap - n_public_segments..final_ap).map(to_memory_value);
+    let mut ranges = start_ptrs
+        .zip(end_ptrs)
+        .map(|(start_ptr, stop_ptr)| SegmentRange {
+            start_ptr,
+            stop_ptr,
+        });
+    let mut present = public_segment_context.into_iter();
+    let mut next = || {
+        let present = present.next().unwrap();
+        if present {
+            ranges.next()
+        } else {
+            None
+        }
+    };
+
+    PublicSegmentRanges {
+        output: next().unwrap(),
+        pedersen: next(),
+        range_check_128: next(),
+        ecdsa: next(),
+        bitwise: next(),
+        ec_op: next(),
+        keccak: next(),
+        poseidon: next(),
+        range_check_96: next(),
+        add_mod: next(),
+        mul_mod: next(),
+    }
+}
+
+pub fn extract_sections_from_memory(
+    memory: &Memory,
+    initial_pc: u32,
+    initial_ap: u32,
+    final_ap: u32,
+    public_segment_context: PublicSegmentContext,
+) -> PublicMemory {
+    let public_segments =
+        extract_public_segments(memory, initial_ap, final_ap, public_segment_context);
+    let program_memory_addresses = initial_pc..initial_ap - 2;
+    let safe_call_addresses = initial_ap - 2..initial_ap;
+    let output_memory_addresses =
+        public_segments.output.start_ptr.value..public_segments.output.stop_ptr.value;
+    let [program, safe_call, output] = [
+        program_memory_addresses,
+        safe_call_addresses,
+        output_memory_addresses,
+    ]
+    .map(|range| {
+        range
+            .map(|addr| {
+                let id = memory.get_raw_id(addr);
+                let value = memory.get(addr).as_u256();
+                (id, value)
+            })
+            .collect_vec()
+    });
+
+    assert!(safe_call.len() == 2);
+
+    assert_eq!(safe_call[0].1, [initial_ap, 0, 0, 0, 0, 0, 0, 0]);
+    assert_eq!(safe_call[1].1, [0, 0, 0, 0, 0, 0, 0, 0]);
+
+    PublicMemory {
+        program,
+        safe_call_ids: array::from_fn(|i| safe_call[i].0),
+        public_segments,
+        output,
+    }
+}
+
+pub fn create_public_data(
+    memory: &Memory,
+    initial_state: CasmState,
+    final_state: CasmState,
+    public_segment_context: PublicSegmentContext,
+) -> PublicData {
+    let initial_pc = initial_state.pc.0;
+    let initial_ap = initial_state.ap.0;
+    let final_ap = final_state.ap.0;
+    let public_memory = extract_sections_from_memory(
+        memory,
+        initial_pc,
+        initial_ap,
+        final_ap,
+        public_segment_context,
+    );
+
+    PublicData {
+        public_memory,
+        initial_state,
+        final_state,
+    }
+}
